+++ src/role_utils.rs
@@ -81,7 +81,7 @@
 //! Each resource can have more operator specific labels.
 
 use std::{
-    collections::{BTreeMap, HashMap},
+    collections::{BTreeMap, BTreeSet, HashMap},
     fmt::{Debug, Display},
 };
 
@@ -93,12 +93,14 @@ use crate::{
     product_config_utils::Configuration,
 };
 use derivative::Derivative;
-use k8s_openapi::apimachinery::pkg::apis::meta::v1::LabelSelector;
+use k8s_openapi::{
+    api::core::v1::PodTemplateSpec, apimachinery::pkg::apis::meta::v1::LabelSelector,
+};
 use kube::{runtime::reflector::ObjectRef, Resource};
-use schemars::JsonSchema;
+use schemars::{schema::Schema, JsonSchema};
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, Default, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
 #[serde(
     rename_all = "camelCase",
     bound(deserialize = "T: Default + Deserialize<'de>")
@@ -117,6 +119,32 @@ pub struct CommonConfiguration<T> {
     // BTreeMap to keep some order with the cli arguments.
     #[serde(default)]
     pub cli_overrides: BTreeMap<String, String>,
+    #[serde(default)]
+    #[schemars(schema_with = "pod_overrides_schema")]
+    pub pod_overrides: PodTemplateSpec,
+}
+
+/// Special schema for PodTemplateSpec without mandatory fields (e.g. `containers`).
+///
+/// The normal PodTemplateSpec requires you to specify `containers` as an `Vec<Container>`.
+/// Often times the user want's to overwrite/add stuff not related to a container
+/// (e.g. tolerations or a ServiceAccount), so it's annoying that he always needs to
+/// specify an empty array for `containers`.
+fn pod_overrides_schema(gen: &mut schemars::gen::SchemaGenerator) -> Schema {
+    let mut schema = PodTemplateSpec::json_schema(gen);
+    if let Schema::Object(ref mut schema_object) = schema {
+        if let Schema::Object(schema_object) = schema_object
+            .object
+            .as_mut()
+            .unwrap()
+            .properties
+            .get_mut("spec")
+            .unwrap()
+        {
+            schema_object.object.as_mut().unwrap().required = BTreeSet::new();
+        }
+    };
+    schema
 }
 
 fn config_schema_default() -> serde_json::Value {
@@ -148,6 +176,7 @@ impl<T: Configuration + 'static> Role<T> {
                 config_overrides: self.config.config_overrides,
                 env_overrides: self.config.env_overrides,
                 cli_overrides: self.config.cli_overrides,
+                pod_overrides: self.config.pod_overrides,
             },
             role_groups: self
                 .role_groups
@@ -162,6 +191,7 @@ impl<T: Configuration + 'static> Role<T> {
                                 config_overrides: group.config.config_overrides,
                                 env_overrides: group.config.env_overrides,
                                 cli_overrides: group.config.cli_overrides,
+                                pod_overrides: group.config.pod_overrides,
                             },
                             replicas: group.replicas,
                             selector: group.selector,