+++ CHANGELOG.md
@@ -12,8 +12,11 @@ All notable changes to this project will be documented in this file.
 
 - Add a new `ResourceRequirementsBuilder` to more easily build resource requirements in a controlled and well defined
   way. ([#598]).
+- Add podOverrides to common struct CommonConfiguration ([#601]).
+- All the operators now must respect the new `podOverrides` attribute! ([#601]).
 
 [#598]: https://github.com/stackabletech/operator-rs/pull/598
+[#601]: https://github.com/stackabletech/operator-rs/pull/601
 
 ## [0.41.0] - 2023-04-20